+++ plotnine/stats/__init__.py
@@ -10,6 +10,7 @@ from .stat_density_2d import stat_density_2d
 from .stat_ecdf import stat_ecdf
 from .stat_ellipse import stat_ellipse
 from .stat_function import stat_function
+from .stat_hull import stat_hull
 from .stat_identity import stat_identity
 from .stat_qq import stat_qq
 from .stat_qq_line import stat_qq_line
@@ -27,7 +28,8 @@ __all__ = ['stat_count', 'stat_bin',
            'stat_bindot', 'stat_boxplot',
            'stat_density', 'stat_ecdf',
            'stat_ellipse', 'stat_density_2d',
-           'stat_function', 'stat_identity',
+           'stat_function', 'stat_hull',
+           'stat_identity',
            'stat_qq', 'stat_qq_line',
            'stat_quantile',
            'stat_smooth', 'stat_sum',